added script for hansel
seperated script for item slot and hansel
