+++ HanselAndGretel/Assets/Scripts/S_GamePlay/DragDrop.cs
@@ -6,7 +6,7 @@ using UnityEngine.EventSystems;
 public class DragDrop : MonoBehaviour, IBeginDragHandler, IEndDragHandler, IDragHandler{
 
     [SerializeField] private Canvas canvas;
-    ItemSlot itemSlot;
+    HanselItemDesc hanselItemDesc;
     private RectTransform rectTransform;
     private CanvasGroup canvasGroup;
 
@@ -14,7 +14,7 @@ public class DragDrop : MonoBehaviour, IBeginDragHandler, IEndDragHandler, IDrag
 
     private void Awake()
     {
-        itemSlot = FindObjectOfType<ItemSlot>();
+        hanselItemDesc = FindObjectOfType<HanselItemDesc>();
         rectTransform = GetComponent<RectTransform>();
         canvasGroup = GetComponent<CanvasGroup>();
     }
@@ -31,8 +31,8 @@ public class DragDrop : MonoBehaviour, IBeginDragHandler, IEndDragHandler, IDrag
         
         canvasGroup.alpha = .6f;
         canvasGroup.blocksRaycasts = false;
-        
-        itemSlot.hideSpeechBubble();
+
+        hanselItemDesc.hideSpeechBubble();
     }
 
     public void OnEndDrag(PointerEventData eventdata)
