+++ HanselAndGretel/Assets/Scripts/S_UI/HanselItemDesc.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f16d5243a6eb7084880d5f442dc3d683
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
