+++ HanselAndGretel/Assets/Scripts/S_UI/ItemSlot.cs
@@ -8,55 +8,21 @@ using UnityEngine.EventSystems;
 public class ItemSlot : MonoBehaviour, IDropHandler
 {
     DragDrop dragDrop;
-    public GameObject speechBubble;
-    public TextMeshProUGUI hanselItemDesc;
+
 
     private bool hasItem;
 
-    public void Awake()
-    {
-        speechBubble.SetActive(false);
-    }
+   
     public void OnDrop(PointerEventData eventData)
     {
-        if (!hasItem && eventData.pointerDrag != null)
+        if (eventData.pointerDrag != null)
         {
 
-            
+
             eventData.pointerDrag.GetComponent<RectTransform>().transform.position = GetComponent<RectTransform>().transform.position;
             dragDrop = eventData.pointerDrag.GetComponent<DragDrop>();
             hasItem = true;
 
-            if (gameObject.CompareTag("Hansel"))
-            {
-                Vector3 yOffset = eventData.pointerDrag.GetComponent<RectTransform>().transform.position;
-                yOffset.y -= 1f;
-                eventData.pointerDrag.GetComponent<RectTransform>().transform.position = yOffset;
-
-                if (hasItem == true)
-                {
-                    showSpeechBubble();
-                }
-
-                else
-                {
-                    hideSpeechBubble();
-                }
-
-            }
-
-
-            
         }
-    }
-    public void hideSpeechBubble()
-    {
-        speechBubble.SetActive(false);
-        hasItem = false;
-    }
-    public void showSpeechBubble()
-    {
-        speechBubble.SetActive(true);
-        hanselItemDesc.text = dragDrop.itemDesc;
     }
 }
